cleanup config package
