+++ server/config/old_config.go
@@ -1,0 +1,10 @@
+package config
+
+type (
+	oldConfig struct {
+		Server    serverConfig    `yaml:",omitempty" mapstructure:"server"`
+		GA        googleAnalytics `yaml:"googleAnalytics,omitempty" mapstructure:"googleAnalytics"`
+		Telemetry telemetry       `yaml:",omitempty" mapstructure:"telemetry"`
+		Demo      demo            `yaml:",omitempty" mapstructure:"demo"`
+	}
+)
